features

+++ src/app/menu/menu-item/menu-item.component.ts
@@ -18,20 +18,25 @@ export class MenuItemComponent {
     private router: Router,
   ) {}
 
-  public onDelete(): void {
-    let ObservableAction
-    ObservableAction = this.menuService.deleteMenu(this.menu)
-    ObservableAction.subscribe({
-      next: menu => {
-        console.log("Suppression OK : ", menu)
-        this.router.navigateByUrl("/").then(
-          ()=>this.router.navigateByUrl("/menus") 
-          )
-      },
-      error: err => {
-        console.log("ERREUR DELETE ", err)
-      }
-    })
+  public onDelete(menu: any): void {
+    // Afficher la boîte de confirmation
+    const confirmation = window.confirm(`Êtes-vous sûr de vouloir supprimer le menu : ${this.menu.nom} ?`);
+    // Si l'utilisateur confirme, procéder à la suppression
+    if (confirmation) {
+      let ObservableAction
+      ObservableAction = this.menuService.deleteMenu(this.menu)
+      ObservableAction.subscribe({
+        next: menu => {
+          console.log("Suppression OK : ", menu)
+          this.router.navigateByUrl("/").then(
+            ()=>this.router.navigateByUrl("/menus") 
+            )
+        },
+        error: err => {
+          console.log("ERREUR DELETE ", err)
+        }
+      })  
+    }
   }
   
 }
